# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management Api Documentation",
        "description": "<b>Api Documentation For</b>, Services responsible for getting the list of all events, get a single event details, creating, updating & deleting an event in database as well as attendee module i.e. getting all attendees for an event, attend an event, get details of an event's attendee & unAttend an Event.",
        "contact": {
            "name": "Aadhar gaur",
            "email": "aadhar41@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://events-management-app.test/api/"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "mante.alicia@example.net"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "object",
                                            "example": "{'token':'15|qw0rsNAs8gn3wUplKsaUBVZhPLUWi4ooRrtZCFVgb65c2164'}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. The provided credentials are incorrect.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Terminate current user login session & logout",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get an User Details",
                "description": "Get an User Details",
                "operationId": "getUserDetails",
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "191"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Darius Lubowitz"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "mante.alicia@example.net"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "id": "517",
                                                "name": "Darius Lubowitz",
                                                "email": "mante.alicia@example.net"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}